# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MultiPanel ERP API",
    "description": "REST API for Plex & Jellyfin management platform",
    "version": "1.0.0",
    "contact": { "name": "MultiPanel Support" }
  },
  "servers": [{ "url": "/api/v1", "description": "API v1" }],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and obtain JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "JWT tokens returned" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
...
    "/dashboard": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Dashboard statistics",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Stats object" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}